refactor: split remaining vsock payload modules
