+++ crates/vsock-proto/src/frame.rs
@@ -98,3 +98,95 @@ impl Default for Decoder {
         Self::new()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::wire::{MSG_PING, MSG_PONG, MSG_READY};
+
+    #[test]
+    fn encode_decode_roundtrip_empty_payload() {
+        let data = encode(MSG_PING, 1, &[]).unwrap();
+        let mut dec = Decoder::new();
+        let msgs = dec.decode(&data).unwrap();
+        assert_eq!(msgs.len(), 1);
+        assert_eq!(msgs[0].msg_type, MSG_PING);
+        assert_eq!(msgs[0].seq, 1);
+        assert!(msgs[0].payload.is_empty());
+    }
+
+    #[test]
+    fn encode_decode_roundtrip_with_payload() {
+        let data = encode(MSG_PING, 42, b"hello world").unwrap();
+        let mut dec = Decoder::new();
+        let msgs = dec.decode(&data).unwrap();
+        assert_eq!(msgs.len(), 1);
+        assert_eq!(msgs[0].msg_type, MSG_PING);
+        assert_eq!(msgs[0].seq, 42);
+        assert_eq!(msgs[0].payload, b"hello world");
+    }
+
+    #[test]
+    fn decoder_handles_partial_reads() {
+        let data = encode(MSG_PONG, 7, &[]).unwrap();
+        let mut dec = Decoder::new();
+
+        // Feed first 4 bytes (header only)
+        let msgs = dec.decode(&data[..4]).unwrap();
+        assert!(msgs.is_empty());
+
+        // Feed the rest
+        let msgs = dec.decode(&data[4..]).unwrap();
+        assert_eq!(msgs.len(), 1);
+        assert_eq!(msgs[0].msg_type, MSG_PONG);
+        assert_eq!(msgs[0].seq, 7);
+    }
+
+    #[test]
+    fn decoder_handles_multiple_messages() {
+        let mut data = encode(MSG_PING, 1, &[]).unwrap();
+        data.extend_from_slice(&encode(MSG_PONG, 1, &[]).unwrap());
+        data.extend_from_slice(&encode(MSG_READY, 0, &[]).unwrap());
+
+        let mut dec = Decoder::new();
+        let msgs = dec.decode(&data).unwrap();
+        assert_eq!(msgs.len(), 3);
+        assert_eq!(msgs[0].msg_type, MSG_PING);
+        assert_eq!(msgs[1].msg_type, MSG_PONG);
+        assert_eq!(msgs[2].msg_type, MSG_READY);
+    }
+
+    #[test]
+    fn decoder_rejects_too_large() {
+        // Craft a header claiming 17MB body
+        let bad = (17 * 1024 * 1024_u32).to_be_bytes();
+        let mut dec = Decoder::new();
+        let err = dec.decode(&bad).unwrap_err();
+        assert!(matches!(err, ProtocolError::MessageTooLarge(_)));
+    }
+
+    #[test]
+    fn decoder_rejects_too_small() {
+        // Body length 2 (less than MIN_BODY_SIZE=5)
+        let bad = 2_u32.to_be_bytes();
+        let mut dec = Decoder::new();
+        let err = dec.decode(&bad).unwrap_err();
+        assert!(matches!(err, ProtocolError::MessageTooSmall(2)));
+    }
+
+    #[test]
+    fn decoder_byte_by_byte() {
+        let data = encode(MSG_PING, 1, &[]).unwrap();
+        let mut dec = Decoder::new();
+
+        for (i, &byte) in data.iter().enumerate() {
+            let msgs = dec.decode(&[byte]).unwrap();
+            if i < data.len() - 1 {
+                assert!(msgs.is_empty());
+            } else {
+                assert_eq!(msgs.len(), 1);
+                assert_eq!(msgs[0].msg_type, MSG_PING);
+            }
+        }
+    }
+}
