+++ crates/vsock-proto/src/payloads/mod.rs
@@ -1,1 +1,5 @@
 pub(crate) mod command;
+pub(crate) mod error;
+pub(crate) mod process;
+pub(crate) mod spawn_watch;
+pub(crate) mod write_file;
